backend/requirements.txt
flask
flask-cors
flask-jwt-extended
openai
faiss-cpu
sentence-transformers
sglang